task 101

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -7,7 +7,7 @@
  * Return: void
  */
 
-void print_number(long int n)
+void print_number(int n)
 {
 	if (n < 0)
 	{
